+++ PowerShellForFOD/Public/Get-FODReleaseAssessmentTypes.ps1
@@ -1,0 +1,145 @@
+function Get-FODReleaseAssessmentTypes {
+    <#
+    .SYNOPSIS
+        Gets the assessment types for a FOD releases.
+    .DESCRIPTION
+        Get the available assessment types for a specific FOD release.
+    .PARAMETER ReleaseId
+        The release id.
+    .PARAMETER ScanType
+        The scan type.
+        Valid values are: Static, Dynamic, Mobile, Monitoring, Network, OpenSource
+    .PARAMETER Filters
+        A delimited list of field filters.
+    .PARAMETER OrderBy
+        The field name to order the results by.
+    .PARAMETER OrderByDirection
+        The direction to order the results by. ASC and DESC are valid values.
+    .PARAMETER Fields
+        Comma separated list of fields to return.
+    .PARAMETER Paging
+        If specified, and more data is available after loading limit of releases,
+        continue querying FOD until we have retrieved all the data available.
+    .PARAMETER Limit
+        Limit the number of releases returned to this number.
+        Maximum value is 50.
+        Default is 50.
+    .PARAMETER Raw
+        If specified, provide raw output and do not parse any responses.
+    .PARAMETER Token
+        FOD token to use.
+        If empty, the value from PS4FOD will be used.
+    .PARAMETER Proxy
+        Proxy server to use.
+        Default value is the value set by Set-FODConfig
+    .EXAMPLE
+        # Get the scans with id 1234 for release id 100 through Paging
+        Get-FODReleaseScan -ReleaseId 100 -ScanId 1234
+    .LINK
+        https://api.ams.fortify.com/swagger/ui/index#!/Releases/ReleasesV3_GetAssessmentTypes
+    .FUNCTIONALITY
+        Fortify on Demand
+    #>
+    [CmdletBinding()]
+    param (
+        [Parameter(Mandatory=$True, ValueFromPipeline=$True)]
+        [int]$ReleaseId,
+
+        [Parameter(Mandatory)]
+        [ValidateNotNullOrEmpty()]
+        [ValidateSet('Static', 'Dynamic', 'Mobile', 'Monitoring', 'Network', 'OpenSource', IgnoreCase = $false)]
+        $ScanType,
+
+        [string]$Filters,
+        [string]$OrderBy,
+        [string]$OrderByDirection,
+        [string]$Fields,
+        [switch]$Raw,
+        [switch]$Paging,
+        [int]$Limit = 50,
+
+        [Parameter()]
+        [ValidateNotNullOrEmpty()]
+        [string]$Token = $Script:PS4FOD.Token,
+
+        [Parameter()]
+        [ValidateNotNullOrEmpty()]
+        [string]$ApiUri = $Script:PS4FOD.ApiUri,
+
+        [Parameter()]
+        [ValidateNotNullOrEmpty()]
+        [string]$Proxy = $Script:PS4FOD.Proxy,
+
+        [switch]$ForceVerbose = $Script:PS4FOD.ForceVerbose
+    )
+    begin
+    {
+        $Params = @{}
+        if ($Proxy) {
+            $Params['Proxy'] = $Proxy
+        }
+        if ($ForceVerbose) {
+            $Params.Add('ForceVerbose', $True)
+            $VerbosePreference = "Continue"
+        }
+        Write-Verbose "Get-FODReleaseAssessmentTypes Bound Parameters: $( $PSBoundParameters | Remove-SensitiveData | Out-String )"
+        $Body = @{
+            scanType = $ScanType
+            offset = 0
+            limit = $Limit
+        }
+        if ($Filters) {
+            $Body.Add("filters", $Filters)
+        }
+        if ($OrderBy) {
+            $Body.Add("orderBy", $OrderBy)
+        }
+        if ($OrderByDirection) {
+            if ($OrderByDirection -eq "ASC" -or $OrderByDirection -eq "DESC") {
+                $Body.Add("orderByDirection", $OrderByDirection)
+            } else {
+                Write-Error "OrderBy can only be ASC or DESC." -ErrorAction Stop
+            }
+        }
+        if ($Fields) {
+            $Body.Add("fields", $Fields)
+        }
+        if ($Limit -gt 50) {
+            Write-Error "Maximum value for Limit is 50." -ErrorAction Stop
+        }
+        $RawAssessmentTypes = @()
+        $HasMore = $false
+        $TotalCount = 0
+        $LoadedCount = 0
+        $LoadLimit = $Limit
+    }
+    process
+    {
+        do {
+            Write-Verbose "Send-FODApi -Method Get -Operation '/api/v3/releases/$ReleaseId/assessment-types'" #$Params
+            $Response = Send-FODApi -Method Get -Operation "/api/v3/releases/$ReleaseId/assessment-types" -Body $Body @Params
+            $TotalCount = $Response.totalCount
+            if ($LoadedCount -lt ($TotalCount - $LoadLimit)) {
+                $HasMore = $true
+                $LoadedCount += $LoadLimit
+                $Body.Remove("offset")
+                $Body.Add("offset", $LoadedCount)
+            } else {
+                $LoadedCount += $TotalCount
+                $HasMore = $false
+            }
+            $RawAssessmentTypes += $Response.items
+        } until (
+        -not $Paging -or
+                -not $HasMore
+        )
+        Write-Verbose "Loaded $LoadedCount release assessment types"
+    }
+    end {
+        if ($Raw) {
+            $RawAssessmentTypes
+        } else {
+            Parse-FODReleaseAssessment -InputObject $RawAssessmentTypes
+        }
+    }
+}